LayoutTests/fast/events/wheel/wheel-event-after-document-open-expected.txt

This tests triggering wheel scroll on an element after document.open. The element should receive wheel event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didFireWheelEvent is true
PASS successfullyParsed is true

TEST COMPLETE
